fix(interview): close finish-screen double-submission window
onFinish did fetch -> router.push and then resolved. App Router
navigation is server-driven, so router.push returns well before the
destination paints. The interview package closes its confirm dialog as
soon as onFinish resolves, which removed the dialog overlay and left the
Finish screen mounted, visible and clickable for the whole navigation.
On a slow connection participants tapped Finish again and submitted a
duplicate finish request.

Swap to the completion screen on client state instead of waiting for
navigation: onFinish now sets  as its last meaningful step,
and the component returns <InterviewCompleted /> when set, unmounting
Shell so the Finish screen cannot be shown or clicked again. React
commits before the next paint, so there is no window to click into.

router.replace is kept but is no longer load-bearing; it only tidies the
URL, and replace rather than push keeps back from returning to the
interview.

A failed request is still allowed to reject so the package keeps the
dialog open and re-enables retry, and an aborted request (dismissing the
dialog while pending) never reaches setFinished, leaving the Finish
screen usable.

The completion screen is extracted from the /interview/finished route
into a shared component so the inline and routed renders cannot drift.
